+++ str.h
@@ -13,20 +13,20 @@
  * terminating '\0' themselves.
  */
 
-#include <stdint.h>
+#include <stdlib.h>
 
 typedef struct str_t_ {
-    char* s;             /* null-terminated string */
-    uint32_t       n;    /* length of s */
-    uint32_t       size; /* bytes allocated for s */
+    char*  s;    /* null-terminated string */
+    size_t n;    /* length of s */
+    size_t size; /* bytes allocated for s */
 } str_t;
 
 str_t* str_new(void);
 void str_free(str_t* str);
 void str_clear(str_t* str);
-void str_reserve(str_t* str, const uint32_t sz);
-void str_extend(str_t* str, const uint32_t ex);
-void str_trunc(str_t* str, const uint32_t tr);
+void str_reserve(str_t* str, const size_t sz);
+void str_extend(str_t* str, const size_t ex);
+void str_trunc(str_t* str, const size_t tr);
 void str_append_str(str_t* str, const str_t* app);
 void str_append_cstr(str_t* str , const char* cstr);
 void str_append_char(str_t* str, char c);
